move remote_storage utility function to S3Storage; wordsmith
